+++ hardware/tests/opentrons_hardware/__init__.py
@@ -1,0 +1,1 @@
+"""Opentrons hardware tests."""
